moved the get_parameter_from_param_file to the cm_services
